requirements.txt
Django==5.2.7
django-allauth==65.11.0
django-ratelimit==4.1.0
python-dotenv==1.1.1
dj-database-url==1.0.0
Pillow==11.3.0
mercadopago==2.3.0
whitenoise[brotli]==6.2.0
gunicorn==20.1.0
PyJWT==2.8.0
isort==7.0.0
psycopg2-binary==2.9.9